fix: 스낵리뷰 유효성 검사 문제 해결 (#183)
* fix: redis base package 경로 변경

* feat: 제품별 스낵리뷰 통계 조회 api 생성

* refactor: 메인페이지 제품조회 메소드 리팩토링 (#181)

* refactor: 스낵리뷰 점수 유효성 검사 안 되는 문제 수정 (#181)

+++ server/src/main/java/seb/project/Codetech/snackreview/validation/validator/ScoreValidator.java
@@ -24,10 +24,10 @@ public class ScoreValidator implements ConstraintValidator<ValidScore, ReviewSco
 				continue;
 			}
 			field.setAccessible(true);
-			int value = 0;
+			float value = 0;
 
 			try {
-				value = field.getInt(score);
+				value = field.getFloat(score);
 			} catch (IllegalAccessException e) {
 			}
